Get new abbreviations

get_new_abbrs() is a Get function that retrieves a pre-existing data object from memory, local file system or online repository. Specifically, this function implements an algorithm to get new abbreviations. Function argument pkg_setup_ls specifies the where to look for the required object. The function returns New abbreviations (a character vector).

get_new_abbrs(
  pkg_setup_ls,
  classes_to_make_tb = NULL,
  inc_all_mthds_1L_lgl = T,
  paths_ls = make_fn_nms(),
  pkg_ds_ls_ls = NULL,
  transformations_chr = NULL,
  undocumented_fns_dir_chr = make_undmtd_fns_dir_chr(drop_empty_1L_lgl = T),
  use_last_1L_int = NULL,
  fns_dmt_tb = deprecated()
)

Arguments

pkg_setup_ls
Package setup (a list)

classes_to_make_tb
Classes to make (a tibble), Default: NULL

inc_all_mthds_1L_lgl
Include all methods (a logical vector of length one), Default: T

paths_ls
Paths (a list), Default: make_fn_nms()

pkg_ds_ls_ls
Package dataset (a list of lists), Default: NULL

transformations_chr
Transformations (a character vector), Default: NULL

undocumented_fns_dir_chr
Undocumented functions directory (a character vector), Default: make_undmtd_fns_dir_chr(drop_empty_1L_lgl = T)

use_last_1L_int
Use last (an integer vector of length one), Default: NULL

fns_dmt_tb
Functions documentation (a tibble), Default: deprecated()

Value

New abbreviations (a character vector)