Make documentation for all functions

make_dmt_for_all_fns() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make documentation for all functions. The function returns All functions documentation (a tibble).

make_dmt_for_all_fns(
  paths_ls = make_fn_nms(),
  undocumented_fns_dir_chr = make_undmtd_fns_dir_chr(),
  custom_dmt_ls = list(details_ls = NULL, inc_for_main_user_lgl_ls =
    list(force_true_chr = NA_character_, force_false_chr = NA_character_), args_ls_ls =
    NULL),
  fn_type_lup_tb,
  abbreviations_lup = NULL,
  object_type_lup = get_rds_from_dv("object_type_lup"),
  inc_all_mthds_1L_lgl = T
)

Arguments

paths_ls
Paths (a list), Default: make_fn_nms()

undocumented_fns_dir_chr
Undocumented functions directory (a character vector), Default: make_undmtd_fns_dir_chr()

custom_dmt_ls
Custom documentation (a list), Default: list(details_ls = NULL, inc_for_main_user_lgl_ls = list(force_true_chr = NA_character_, force_false_chr = NA_character_), args_ls_ls = NULL)

fn_type_lup_tb
Function type lookup table (a tibble)

abbreviations_lup
Abbreviations (a lookup table), Default: NULL

object_type_lup
Object type (a lookup table), Default: get_rds_from_dv("object_type_lup")

inc_all_mthds_1L_lgl
Include all methods (a logical vector of length one), Default: T

Value

All functions documentation (a tibble)