Make function documentation table template

make_fn_dmt_tbl_tmpl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make function documentation table template. The function returns Function documentation table (a tibble).

make_fn_dmt_tbl_tmpl(
  fns_path_chr,
  fns_dir_chr = make_undmtd_fns_dir_chr(drop_empty_1L_lgl = T),
  fns_env_ls = NULL,
  fn_types_lup = NULL,
  abbreviations_lup = NULL,
  dv_ds_nm_1L_chr = "ready4-dev/ready4",
  dv_url_pfx_1L_chr = deprecated(),
  key_1L_chr = deprecated(),
  object_type_lup = NULL,
  server_1L_chr = deprecated(),
  test_for_write_R_warning_fn = NULL
)

Arguments

fns_path_chr

Functions path (a character vector)

fns_dir_chr

Functions directory (a character vector), Default: make_undmtd_fns_dir_chr(drop_empty_1L_lgl = T)

fns_env_ls

Functions (a list of environments), Default: NULL

fn_types_lup

Function types (a lookup table), Default: NULL

abbreviations_lup

Abbreviations (a lookup table), Default: NULL

dv_ds_nm_1L_chr

Dataverse dataset name (a character vector of length one), Default: 'ready4-dev/ready4'

dv_url_pfx_1L_chr

Dataverse url prefix (a character vector of length one), Default: deprecated()

key_1L_chr

Key (a character vector of length one), Default: deprecated()

object_type_lup

Object type (a lookup table), Default: NULL

server_1L_chr

Server (a character vector of length one), Default: deprecated()

test_for_write_R_warning_fn

Test for write warning (a function), Default: NULL

Value

Function documentation table (a tibble)