Make function documentation table template

make_fn_dmt_tbl_tpl() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make function documentation table template. The function returns Function documentation table (a tibble).

make_fn_dmt_tbl_tpl(
  fns_path_chr,
  fns_dir_chr = make_undmtd_fns_dir_chr(),
  fn_type_lup_tb = NULL,
  abbreviations_lup = NULL,
  object_type_lup = NULL,
  test_for_write_R_warning_fn = NULL
)

Arguments

fns_path_chr
Functions path (a character vector)

fns_dir_chr
Functions directory (a character vector), Default: make_undmtd_fns_dir_chr()

fn_type_lup_tb
Function type lookup table (a tibble), Default: NULL

abbreviations_lup
Abbreviations (a lookup table), Default: NULL

object_type_lup
Object type (a lookup table), Default: NULL

test_for_write_R_warning_fn
Test for write warning (a function), Default: NULL

Value

Function documentation table (a tibble)