Make manifest

make_manifest() is a Make function that creates a new R object. Specifically, this function implements an algorithm to make manifest. The function returns Manifest (a list).

make_manifest(
  pkg_desc_ls,
  copyright_holders_chr,
  pkg_dmt_dv_dss_chr = deprecated(),
  add_gh_site_1L_lgl = T,
  addl_badges_ls = list(),
  addl_pkgs_ls = make_addl_pkgs_ls(),
  badges_lup = tibble::tibble(),
  build_ignore_ls = make_build_ignore_ls(),
  check_type_1L_chr = "ready4",
  classify_1L_lgl = T,
  cls_fn_ls = list(),
  custom_dmt_ls = make_custom_dmt_ls(),
  delete_r_dir_cnts_1L_lgl = T,
  dev_pkg_nm_1L_chr = get_dev_pkg_nm(getwd()),
  dev_pkgs_chr = NA_character_,
  dss_records_ls = list(),
  dv_url_pfx_1L_chr = character(0),
  gh_repo_1L_chr = NA_character_,
  import_from_chr = character(0),
  lifecycle_stage_1L_chr = "experimental",
  inc_pkg_meta_data_1L_lgl = F,
  incr_ver_1L_lgl = F,
  key_1L_chr = NULL,
  on_cran_1L_lgl = F,
  path_to_dmt_dir_1L_chr = normalizePath("../../../../../Documentation/Code"),
  path_to_pkg_logo_1L_chr = NA_character_,
  path_to_pkg_rt_1L_chr = getwd(),
  piggyback_to_1L_chr = NA_character_,
  pkg_ds_ls_ls = list(),
  ready4_type_1L_chr,
  s4_mthds_fn = NULL,
  s4_mthds_ls = NULL,
  server_1L_chr = Sys.getenv("DATAVERSE_SERVER"),
  user_manual_fns_chr = NA_character_,
  zenodo_badge_1L_chr = character(0)
)

Arguments

pkg_desc_ls

Package description (a list)

copyright_holders_chr

Copyright holders (a character vector)

pkg_dmt_dv_dss_chr

Package documentation dataverse datasets (a character vector), Default: deprecated()

add_gh_site_1L_lgl

Add github site (a logical vector of length one), Default: T

addl_badges_ls

Additional badges (a list), Default: list()

addl_pkgs_ls

Additional packages (a list), Default: make_addl_pkgs_ls()

badges_lup

Badges (a lookup table), Default: tibble::tibble()

build_ignore_ls

Build ignore (a list), Default: make_build_ignore_ls()

check_type_1L_chr

Check type (a character vector of length one), Default: 'ready4'

classify_1L_lgl

Classify (a logical vector of length one), Default: T

cls_fn_ls

Class (a list of functions), Default: list()

custom_dmt_ls

Custom documentation (a list), Default: make_custom_dmt_ls()

delete_r_dir_cnts_1L_lgl

Delete r directory contents (a logical vector of length one), Default: T

dev_pkg_nm_1L_chr

Development package name (a character vector of length one), Default: get_dev_pkg_nm(getwd())

dev_pkgs_chr

Development packages (a character vector), Default: 'NA'

dss_records_ls

Datasets records (a list), Default: list()

dv_url_pfx_1L_chr

Dataverse url prefix (a character vector of length one), Default: character(0)

gh_repo_1L_chr

Github repository (a character vector of length one), Default: 'NA'

import_from_chr

Import from (a character vector), Default: character(0)

lifecycle_stage_1L_chr

Lifecycle stage (a character vector of length one), Default: 'experimental'

inc_pkg_meta_data_1L_lgl

Include package meta data (a logical vector of length one), Default: F

incr_ver_1L_lgl

Increment version (a logical vector of length one), Default: F

key_1L_chr

Key (a character vector of length one), Default: NULL

on_cran_1L_lgl

On cran (a logical vector of length one), Default: F

path_to_dmt_dir_1L_chr

Path to documentation directory (a character vector of length one), Default: normalizePath("../../../../../Documentation/Code")

path_to_pkg_logo_1L_chr

Path to package logo (a character vector of length one), Default: 'NA'

path_to_pkg_rt_1L_chr

Path to package root (a character vector of length one), Default: getwd()

piggyback_to_1L_chr

Piggyback to (a character vector of length one), Default: 'NA'

pkg_ds_ls_ls

Package dataset (a list of lists), Default: list()

ready4_type_1L_chr

Ready4 type (a character vector of length one)

s4_mthds_fn

S4 methods (a function), Default: NULL

s4_mthds_ls

S4 methods (a list), Default: NULL

server_1L_chr

Server (a character vector of length one), Default: Sys.getenv("DATAVERSE_SERVER")

user_manual_fns_chr

User manual functions (a character vector), Default: 'NA'

zenodo_badge_1L_chr

Zenodo badge (a character vector of length one), Default: character(0)

Value

Manifest (a list)