Renew values in a dataset

renew.ready4fun_abbreviations() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the ready4 submodule class for tibble object lookup table of abbreviations. The function is called for its side effects and does not return a value.

renew.ready4fun_functions() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the Function types lookup table. The function is called for its side effects and does not return a value.

renew.ready4fun_manifest() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the ready4 submodule class for encapsulating the metadata required for package set-up. The function is called for its side effects and does not return a value.

renew.ready4fun_objects() is a renew method that renews an instance of a class by updating it with new data. This method is implemented for the Object types lookup table. The function is called for its side effects and does not return a value.

Usage

renew.ready4fun_abbreviations(
  x,
  short_name_chr = NA_character_,
  long_name_chr = NA_character_,
  plural_lgl = NA,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  slice_indcs_int = NA_integer_
)

# S4 method for ready4fun_abbreviations
renew(
  x,
  short_name_chr = NA_character_,
  long_name_chr = NA_character_,
  plural_lgl = NA,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  slice_indcs_int = NA_integer_
)

renew.ready4fun_functions(
  x,
  fn_type_nm_chr = NA_character_,
  fn_type_desc_chr = NA_character_,
  first_arg_desc_chr = NA_character_,
  second_arg_desc_chr = NA_character_,
  is_generic_lgl = NA,
  is_method_lgl = NA,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  slice_indcs_int = NA_integer_
)

# S4 method for ready4fun_functions
renew(
  x,
  fn_type_nm_chr = NA_character_,
  fn_type_desc_chr = NA_character_,
  first_arg_desc_chr = NA_character_,
  second_arg_desc_chr = NA_character_,
  is_generic_lgl = NA,
  is_method_lgl = NA,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  slice_indcs_int = NA_integer_
)

renew.ready4fun_manifest(
  x,
  type_1L_chr,
  append_1L_lgl = F,
  are_words_chr = character(0),
  custom_plural_ls = NULL,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  long_name_chr = character(0),
  no_plural_chr = character(0),
  not_obj_type_chr = character(0),
  pfx_rgx = character(0),
  tf_to_singular_chr = character(0)
)

# S4 method for ready4fun_manifest
renew(
  x,
  type_1L_chr,
  append_1L_lgl = F,
  are_words_chr = character(0),
  custom_plural_ls = NULL,
  key_1L_chr = Sys.getenv("DATAVERSE_KEY"),
  long_name_chr = character(0),
  no_plural_chr = character(0),
  not_obj_type_chr = character(0),
  pfx_rgx = character(0),
  tf_to_singular_chr = character(0)
)

renew.ready4fun_objects(
  x,
  short_name_chr = NA_character_,
  long_name_chr = NA_character_,
  atomic_element_lgl = NA,
  r3_can_extend_lgl = NA,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  slice_indcs_int = NA_integer_
)

# S4 method for ready4fun_objects
renew(
  x,
  short_name_chr = NA_character_,
  long_name_chr = NA_character_,
  atomic_element_lgl = NA,
  r3_can_extend_lgl = NA,
  filter_cdn_1L_chr = NA_character_,
  new_cases_r3 = NULL,
  slice_indcs_int = NA_integer_
)

Arguments

x

An instance of ready4fun_objects, a Object types lookup table.

short_name_chr

Short name (a character vector), Default: 'NA'

long_name_chr

Long name (a character vector), Default: 'NA'

plural_lgl

Plural (a logical vector), Default: NA

filter_cdn_1L_chr

Filter condition (a character vector of length one), Default: 'NA'

new_cases_r3

New cases (a ready4 submodule), Default: NULL

slice_indcs_int

Slice indices (an integer vector), Default: NA

fn_type_nm_chr

Function type name (a character vector), Default: 'NA'

fn_type_desc_chr

Function type description (a character vector), Default: 'NA'

first_arg_desc_chr

First argument description (a character vector), Default: 'NA'

second_arg_desc_chr

Second argument description (a character vector), Default: 'NA'

is_generic_lgl

Is generic (a logical vector), Default: NA

is_method_lgl

Is method (a logical vector), Default: NA

type_1L_chr

Type (a character vector of length one)

append_1L_lgl

Append (a logical vector of length one), Default: F

are_words_chr

Are words (a character vector), Default: character(0)

custom_plural_ls

Custom plural (a list), Default: NULL

key_1L_chr

Key (a character vector of length one), Default: Sys.getenv("DATAVERSE_KEY")

no_plural_chr

No plural (a character vector), Default: character(0)

not_obj_type_chr

Not object type (a character vector), Default: character(0)

pfx_rgx

Prefix (a regular expression vector), Default: character(0)

tf_to_singular_chr

Transform to singular (a character vector), Default: character(0)

atomic_element_lgl

Atomic element (a logical vector), Default: NA

r3_can_extend_lgl

Ready4 submodule can extend (a logical vector), Default: NA

Value

x (An object)

x (An object)

x (An object)

x (An object)