Update functions documentation tibble

update_fns_dmt_tb() is an Update function that edits an object, while preserving core object attributes. Specifically, this function implements an algorithm to update functions documentation tibble. Function argument fns_dmt_tb specifies the object to be updated. Argument title_ls provides the object to be updated. The function returns Functions documentation (a tibble).

Usage,

update_fns_dmt_tb(
  fns_dmt_tb,
  title_ls = NULL,
  desc_ls = NULL,
  details_ls = NULL,
  inc_for_main_user_lgl_ls = NULL,
  output_ls = NULL,
  example_ls = NULL,
  args_ls_ls = NULL,
  append_1L_lgl = T
)

Arguments

fns_dmt_tb

Functions documentation (a tibble)

title_ls

Title (a list), Default: NULL

desc_ls

Description (a list), Default: NULL

details_ls

Details (a list), Default: NULL

inc_for_main_user_lgl_ls

Include for main user (a list of logical vectors), Default: NULL

output_ls

Output (a list), Default: NULL

example_ls

Example (a list), Default: NULL

args_ls_ls

Arguments (a list of lists), Default: NULL

append_1L_lgl

Append (a logical vector of length one), Default: T

Value

Functions documentation (a tibble)