write abbreviation lookup table

write_abbr_lup() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write abbreviation lookup table. The function returns Package datasets (a tibble).

write_abbr_lup(
  seed_lup = NULL,
  short_name_chr = NA_character_,
  long_name_chr = NA_character_,
  no_plural_chr = NA_character_,
  custom_plural_ls = NULL,
  overwrite_1L_lgl = T,
  object_type_lup = NULL,
  pkg_dss_tb = tibble::tibble(ds_obj_nm_chr = character(0), title_chr = character(0),
    desc_chr = character(0), url_chr = character(0)),
  pkg_nm_1L_chr = get_dev_pkg_nm(),
  dv_ds_nm_1L_chr = "ready4-dev/ready4",
  dv_url_pfx_1L_chr = deprecated(),
  key_1L_chr = deprecated(),
  server_1L_chr = deprecated(),
  url_1L_chr = deprecated()
)

Arguments

seed_lup
Seed (a lookup table), Default: NULL

short_name_chr
Short name (a character vector), Default: 'NA'

long_name_chr
Long name (a character vector), Default: 'NA'

no_plural_chr
No plural (a character vector), Default: 'NA'

custom_plural_ls
Custom plural (a list), Default: NULL

overwrite_1L_lgl
Overwrite (a logical vector of length one), Default: T

object_type_lup
Object type (a lookup table), Default: NULL

pkg_dss_tb
Package datasets (a tibble), Default: tibble::tibble(ds_obj_nm_chr = character(0), title_chr = character(0), desc_chr = character(0), url_chr = character(0))

pkg_nm_1L_chr
Package name (a character vector of length one), Default: get_dev_pkg_nm()

dv_ds_nm_1L_chr
Dataverse dataset name (a character vector of length one), Default: 'ready4-dev/ready4'

dv_url_pfx_1L_chr
Dataverse url prefix (a character vector of length one), Default: deprecated()

key_1L_chr
Key (a character vector of length one), Default: deprecated()

server_1L_chr
Server (a character vector of length one), Default: deprecated()

url_1L_chr
Url (a character vector of length one), Default: deprecated()

Value

Package datasets (a tibble)