Write new object types

write_new_obj_types() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write new object types. The function returns Package setup (a list).

write_new_obj_types(
  pkg_setup_ls,
  long_name_chr = NULL,
  atomic_element_lgl = F,
  custom_plural_ls = NULL,
  key_1L_chr = deprecated(),
  no_plural_chr = NA_character_,
  publish_dv_1L_lgl = deprecated(),
  pfx_rgx = NA_character_,
  r3_can_extend_lgl = F,
  server_1L_chr = deprecated()
)

Arguments

pkg_setup_ls

Package setup (a list)

long_name_chr

Long name (a character vector), Default: NULL

atomic_element_lgl

Atomic element (a logical vector), Default: F

custom_plural_ls

Custom plural (a list), Default: NULL

key_1L_chr

Key (a character vector of length one), Default: deprecated()

no_plural_chr

No plural (a character vector), Default: 'NA'

publish_dv_1L_lgl

Publish dataverse (a logical vector of length one), Default: deprecated()

pfx_rgx

Prefix (a regular expression vector), Default: 'NA'

r3_can_extend_lgl

Ready4 S3 can extend (a logical vector), Default: F

server_1L_chr

Server (a character vector of length one), Default: deprecated()

Value

Package setup (a list)