Write package setup files

write_pkg_setup_fls() is a Write function that writes a file to a specified local directory. Specifically, this function implements an algorithm to write package setup files. The function is called for its side effects and does not return a value. WARNING: This function writes R scripts to your local environment. Make sure to only use if you want this behaviour

write_pkg_setup_fls(
  pkg_desc_ls,
  path_to_pkg_rt_1L_chr = getwd(),
  dev_pkg_nm_1L_chr = get_dev_pkg_nm(getwd()),
  incr_ver_1L_lgl = T,
  delete_r_dir_cnts_1L_lgl = F,
  copyright_holders_chr,
  check_type_1L_chr = "none",
  add_gh_site_1L_lgl = T,
  path_to_pkg_logo_1L_chr = NA_character_,
  github_repo_1L_chr,
  lifecycle_stage_1L_chr = "experimental",
  badges_lup = NULL,
  addl_badges_ls = NULL
)

Arguments

pkg_desc_ls
Package description (a list)

path_to_pkg_rt_1L_chr
Path to package root (a character vector of length one), Default: getwd()

dev_pkg_nm_1L_chr
Development package name (a character vector of length one), Default: get_dev_pkg_nm(getwd())

incr_ver_1L_lgl
Incr ver (a logical vector of length one), Default: T

delete_r_dir_cnts_1L_lgl
Delete r directory contents (a logical vector of length one), Default: F

copyright_holders_chr
Copyright holders (a character vector)

check_type_1L_chr
Check type (a character vector of length one), Default: 'none'

add_gh_site_1L_lgl
Add gh site (a logical vector of length one), Default: T

path_to_pkg_logo_1L_chr
Path to package logo (a character vector of length one), Default: 'NA'

github_repo_1L_chr
Github repo (a character vector of length one)

lifecycle_stage_1L_chr
Lifecycle stage (a character vector of length one), Default: 'experimental'

badges_lup
Badges (a lookup table), Default: NULL

addl_badges_ls
Addl badges (a list), Default: NULL